feat: add domain crate and move epoch_snapshot into it
