+++ crates/actors/src/block_discovery.rs
@@ -1,5 +1,4 @@
 use crate::{
-    block_index_service::BlockIndexReadGuard,
     block_tree_service::{BlockTreeReadGuard, BlockTreeServiceMessage},
     block_validation::prevalidate_block,
     mempool_service::MempoolServiceMessage,
@@ -12,6 +11,7 @@ use irys_database::{
     block_header_by_hash, commitment_tx_by_txid, db::IrysDatabaseExt as _, tx_header_by_txid,
     CommitmentSnapshotStatus, SystemLedger,
 };
+use irys_domain::block_index_guard::BlockIndexReadGuard;
 use irys_reward_curve::HalvingCurve;
 use irys_types::{
     BlockHash, CommitmentTransaction, Config, DataLedger, DatabaseProvider, GossipBroadcastMessage,
