+++ crates/actors/src/block_producer.rs
@@ -22,6 +22,7 @@ use alloy_signer_local::LocalSigner;
 use base58::ToBase58 as _;
 use eyre::eyre;
 use irys_database::{block_header_by_hash, db::IrysDatabaseExt as _, SystemLedger};
+use irys_domain::block_index_guard::BlockIndexReadGuard;
 use irys_price_oracle::IrysPriceOracle;
 use irys_reth::compose_shadow_tx;
 use irys_reth_node_bridge::IrysRethNodeAdapter;
@@ -49,7 +50,6 @@ use tokio::sync::oneshot;
 use tracing::{debug, error, info, warn, Instrument as _, Span};
 
 mod block_validation_tracker;
-use crate::block_index_service::BlockIndexReadGuard;
 pub use block_validation_tracker::BlockValidationTracker;
 use irys_types::block_provider::ResetSeedCache;
