+++ crates/actors/src/block_validation.rs
@@ -2,13 +2,11 @@ use std::{sync::Arc, time::Duration};
 
 use crate::{
     block_discovery::{get_commitment_tx_in_parallel, get_data_tx_in_parallel},
-    block_index_service::BlockIndexReadGuard,
     block_tree_service::ema_snapshot::EmaSnapshot,
     mempool_service::MempoolServiceMessage,
     mining::hash_to_number,
     services::ServiceSenders,
     shadow_tx_generator::ShadowTxGenerator,
-    EpochSnapshot,
 };
 use alloy_consensus::Transaction as _;
 use alloy_eips::eip7685::{Requests, RequestsOrHash};
@@ -17,6 +15,7 @@ use async_trait::async_trait;
 use base58::ToBase58 as _;
 use eyre::{ensure, OptionExt as _};
 use irys_database::{block_header_by_hash, db::IrysDatabaseExt as _, SystemLedger};
+use irys_domain::{BlockIndexReadGuard, EpochSnapshot};
 use irys_packing::{capacity_single::compute_entropy_chunk, xor_vec_u8_arrays_in_place};
 use irys_reth::alloy_rlp::Decodable as _;
 use irys_reth::shadow_tx::ShadowTransaction;
@@ -690,13 +689,12 @@ fn validate_shadow_transactions_match(
 mod tests {
     use crate::{
         block_index_service::{BlockIndexService, GetBlockIndexGuardMessage},
-        epoch_service::EpochSnapshot,
         BlockFinalizedMessage,
     };
     use actix::{prelude::*, SystemRegistry};
-
     use irys_config::StorageSubmodulesConfig;
     use irys_database::{add_genesis_commitments, BlockIndex};
+    use irys_domain::{BlockIndexReadGuard, EpochSnapshot};
     use irys_testing_utils::utils::temporary_directory;
     use irys_types::{
         irys::IrysSigner, partition::PartitionAssignment, Address, Base64, DataTransactionLedger,
