+++ crates/api-server/Cargo.toml
@@ -7,6 +7,7 @@ version = "0.1.0"
 # Irys
 irys-actors.workspace = true
 irys-database.workspace = true
+irys-domain.workspace = true
 irys-reth-node-bridge.workspace = true
 irys-storage.workspace = true
 irys-types.workspace = true
