+++ crates/chain/tests/api/tx_commitments.rs
@@ -2,9 +2,10 @@ use crate::utils::*;
 use assert_matches::assert_matches;
 use base58::ToBase58 as _;
 use eyre::eyre;
-use irys_actors::{packing::wait_for_packing, EpochSnapshot};
+use irys_actors::packing::wait_for_packing;
 use irys_chain::IrysNodeCtx;
 use irys_database::CommitmentSnapshotStatus;
+use irys_domain::EpochSnapshot;
 use irys_primitives::CommitmentType;
 use irys_testing_utils::initialize_tracing;
 use irys_types::{irys::IrysSigner, Address, CommitmentTransaction, NodeConfig, H256};
