+++ crates/domain/src/epoch_snapshot/mod.rs
@@ -1,11 +1,9 @@
 pub mod commitment_state;
 pub mod epoch_replay_data;
-pub mod epoch_service;
-pub mod epoch_service_messages;
+pub mod epoch_snapshot;
 pub mod partition_assignments;
 
 pub use commitment_state::*;
 pub use epoch_replay_data::*;
-pub use epoch_service::*;
-pub use epoch_service_messages::*;
+pub use epoch_snapshot::*;
 pub use partition_assignments::*;
