+++ crates/domain/src/epoch_snapshot/partition_assignments.rs
