+++ crates/domain/src/guards/mod.rs
@@ -1,0 +1,2 @@
+pub mod block_index_guard;
+pub use block_index_guard::*;
