+++ crates/domain/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod epoch_snapshot;
+pub mod guards;
+
+pub use epoch_snapshot::*;
+pub use guards::*;
